requirements.txt
# Core dependencies for RAG-based research assistant



numpy
faiss-cpu
sentence-transformers
transformers
torch
pypdf
groq
